Finalizado entrada e saida, test

+++ backend-EasyPark/src/main/java/com/backend/EasyPark/service/AssinaturaPlanoService.java
@@ -5,6 +5,7 @@ import com.backend.EasyPark.entities.AssinaturaPlano;
 import com.backend.EasyPark.entities.Plano;
 import com.backend.EasyPark.entities.Usuario;
 import com.backend.EasyPark.enums.TipoPlano;
+import com.backend.EasyPark.exception.EstacionamentoException;
 import com.backend.EasyPark.repository.AssinaturaPlanoRepository;
 import com.backend.EasyPark.repository.PlanoRepository;
 import com.backend.EasyPark.repository.UsuarioRepository;
@@ -16,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -37,31 +39,38 @@ public class AssinaturaPlanoService {
     }
 
     @Transactional
-    public AssinaturaPlanoDTO create(AssinaturaPlanoDTO dto) {
-        // Verifica se o usuário já possui uma assinatura ativa
+    public AssinaturaPlanoDTO create(AssinaturaPlanoDTO dto) throws EstacionamentoException {
+
+        //VERIFICAR UMA ASSINATURA ATIVA PARA O USUARIO
+        if (assinaturaPlanoRepository.existsByUsuarioIdAndPlanoId(dto.getUsuarioDTO().getId(), dto.getPlanoDTO().getId())) {
+            throw new EstacionamentoException("O usuário já possui uma assinatura ativa para este plano.");
+        }
+
+        //VARIAVEL PARA OBTER INFORMAÇOES
         AssinaturaPlano entity = AssinaturaPlanoMapper.toEntity(dto);
-        // Busca e valida o usuário
+
+        //Busca e valida o usuário
         Usuario usuario = usuarioRepository.findById(dto.getUsuarioDTO().getId())
-                .orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado"));
+                .orElseThrow(() -> new EstacionamentoException("Usuário não encontrado"));
         entity.setUsuario(usuario);
 
-        // Busca e valida o plano
+        //Busca e valida o plano
         Plano plano = planoRepository.findById(dto.getPlanoDTO().getId())
                 .orElseThrow(() -> new EntityNotFoundException("Plano não encontrado"));
         entity.setPlano(plano);
 
-        // Define a data atual como data de pagamento
+        //Define a data atual como data de pagamento
         LocalDateTime dataPagamento = LocalDateTime.now();
         entity.setDataPagamento(dataPagamento);
 
-        // Define a data de vencimento como 30 dias após o pagamento
+        //Define a data de vencimento como 30 dias após o pagamento
         LocalDateTime dataVencimento = dataPagamento.plusDays(30);
         entity.setDataVencimento(dataVencimento);
 
-        // Marca a assinatura como ativa
+        //Marca a assinatura como ativa
         entity.setAtivo(true);
 
-        // Valida o tipo do plano
+        //Valida o tipo do plano
         if (!isTipoPlanoValido(plano.getTipoPlano())) {
             throw new IllegalArgumentException("Tipo de plano inválido. Deve ser INTEGRAL, MANHA ou TARDE");
         }
@@ -72,19 +81,10 @@ public class AssinaturaPlanoService {
     private boolean isTipoPlanoValido(TipoPlano tipoPlano) {
         return tipoPlano == TipoPlano.INTEGRAL ||
                 tipoPlano == TipoPlano.MANHA ||
-                tipoPlano == TipoPlano.TARDE;
+                tipoPlano == TipoPlano.TARDE ||
+                tipoPlano == TipoPlano.NOITE;
     }
 
-    // Método para verificar se o horário atual está dentro do período do plano
-    public boolean isHorarioPermitido(TipoPlano tipoPlano) {
-        int horaAtual = LocalDateTime.now().getHour();
-
-        return switch (tipoPlano) {
-            case INTEGRAL -> true; // Pode acessar em qualquer horário
-            case MANHA -> horaAtual >= 6 && horaAtual < 12; // 6h às 12h
-            case TARDE -> horaAtual >= 12 && horaAtual < 18; // 12h às 18h
-        };
-    }
 
     @Transactional
     public AssinaturaPlanoDTO update(Integer id, AssinaturaPlanoDTO dto) {
@@ -109,10 +109,16 @@ public class AssinaturaPlanoService {
     }
 
     @Transactional
-    public void delete(Integer id) {
+    public void delete(Integer id) throws EstacionamentoException {
         if (!assinaturaPlanoRepository.existsById(id)) {
             throw new EntityNotFoundException("Assinatura não encontrada");
         }
+
+        Optional<AssinaturaPlano> assinaturaPlano = assinaturaPlanoRepository.findById(id);
+        if (assinaturaPlano.isPresent() && assinaturaPlano.get().isAtivo()) {
+            throw new EstacionamentoException("Nao pode excluir assinatura antes de desativala");
+        }
+
         assinaturaPlanoRepository.deleteById(id);
     }
 }
